renamed ui -> view, updated dependecies
