+++ diode_measurement/view/general.py
