+++ diode_measurement/view/logwindow.py
