+++ diode_measurement/view/mainwindow.py
