+++ diode_measurement/view/panels.py
