+++ diode_measurement/view/resource.py
