+++ diode_measurement/view/role.py
