+++ diode_measurement/view/widgets.py
