+++ requirements.txt
@@ -2,5 +2,5 @@ PyQt5==5.15.*
 PyQtChart==5.15.*
 PyVISA==1.11.*
 PyVISA-py==0.5.*
-pint==0.17.*
+pint==0.19.*
 json-rpc==1.13.*
