+++ setup.cfg
@@ -20,7 +20,7 @@ install_requires =
     PyQtChart==5.15.*
     PyVISA==1.11.*
     PyVISA-py==0.5.*
-    pint==0.17.*
+    pint==0.19.*
     json-rpc==1.13.*
 include_package_data = True
