+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py37, py38, py39, py310
+envlist = py38, py39, py310, py311
 isolated_build = true
 skip_missing_interpreters = true
